Fixed memory from 64-bit to 32-bit

+++ pkg/util/util.go
@@ -5,13 +5,17 @@ import (
 	"strings"
 )
 
-func ConvertToHex(num int64, minLength int) string {
+// Creates a string representation of a number in hexadecimal format
+func ConvertToHex(num int32, minLength int) string {
+	// Convert the number to hex
 	converted := fmt.Sprintf("%x", num)
+	// Add leading 0s if needed
 	if len(converted) < minLength {
 		diff := minLength - len(converted)
 		pad := strings.Repeat("0", diff)
 		converted = pad + converted
 	}
+	// Add the '0x' prefix and make it all uppercase
 	converted = "0x" + strings.ToUpper(converted)
 	return converted
 }
